requirements.txt
tensorflow
opencv-python
tkinter
Pillow
numpy
mysql-connector-python

